VmdPrototype
